+++ src/lib/logger.js
@@ -1,23 +1,30 @@
-/* eslint no-console: 0 */
-/* eslint no-param-reassign: 0 */
-
 import bunyan from 'bunyan';
 import bformat from 'bunyan-format';
 import _ from 'lodash';
 
+
 const formatOut = bformat({ outputMode: 'short' });
 
+export const initLogger = () => {
+  const logger = bunyan.createLogger({
+    name: 'server',
+    stream: formatOut,
+    level: 'debug',
+  });
 
-export const log = console.log;
+  // use the logger for console output too
+  logger.console = console.log;
+  console.log = logger.info.bind(logger);
+  console.info = logger.info.bind(logger);
+  console.debug = logger.debug.bind(logger);
+  console.error = logger.error.bind(logger);
+  console.warn = logger.warn.bind(logger);
 
-export const logger = bunyan.createLogger({
-  name: 'server',
-  stream: formatOut,
-  level: 'debug',
-});
+  return logger;
+};
 
 /**
- * Flush logs and display them via logger.
+ * Flush logs and display them via console.
  */
 export const displayLogs = logs => {
   if (logs === undefined) {
@@ -25,14 +32,14 @@ export const displayLogs = logs => {
   }
 
   if (_.isString(logs)) {
-    logs = [ logs ];
+    logs = [logs];    // eslint-disable-line no-param-reassign
   }
 
-  logs = logs.reverse();
+  logs = logs.reverse();  // eslint-disable-line no-param-reassign
 
   if (_.isArray(logs)) {
     while (logs.length) {
-      logger.info(` => ${logs.pop()}`);
+      console.info(` => ${logs.pop()}`);
     }
   }
 };