+++ src/main.js
@@ -4,5 +4,5 @@ export { combineReducers } from 'redux';
 export SchemaObject from 'node-schema-object';
 
 export { api } from './api';
-export { combineFixtures, combineFixturesRoutes, combineFixturesReducers, }
+export { combineFixtures, combineFixturesRoutes, combineFixturesReducers }
   from './lib';